Time: 78 ms (68.71%), Space: 10.5 MB (62.58%) - LeetHub

+++ 923-3sum-with-multiplicity/923-3sum-with-multiplicity.cpp
@@ -50,7 +50,7 @@ public:
                 }
             }
         }
-        return (int) ans;
+        return ans;
         
         // return count;
